# backend/swagger.yaml
swagger: "2.0"
info:
  description: "APIs of Logistic Web Service"
  version: "1.0.0"
  title: "Logistic Web Service"
  contact:
    email: "d.bed11@gptgdt.net"
host: "localhost:8000"
basePath: "/api"
schemes:
- "http"
- "https"
...
paths:
...
  /prohibited-product:
    get:
      tags:
        - "prohibited-product"
      summary: get prohibited products
      produces:
        - application/json
      parameters:
        - in: query
          name: keyword
          type: string
          description: keyword of prohibited product 
        - in: query
          name: page
          type: string
          description: number of page 
        - in: query
          name: pageSize
          type: string
          description: number of page size 
      responses:
        200:
          description: get prohibited product successfully.
          schema:
            type: object
            example: {
                "name": "string",
                "detail": "string",
                "image": "path",
            }
        400:
          description: Prohibited product in not found.    
  /prohibited-product/{id}:
    get:
      tags:
        - "prohibited-product"
      summary: get prohibited product by id
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: Id of a prohibited product 
      responses:
        200:
          description: get prohibited product successfully.
          schema:
            type: object
            example: {
                "name": "string",
                "detail": "string",
                "image": "path",
            }
        400:
          description: Id is not found. 
...
  /admin/prohibited-product/{id}:
    put:
      tags:
        - "prohibited-product"
      summary: "Update a existing prohibited product"
      consumes:
        - application/json
      parameters:
        - in: path
          name: id
          schema: 
            type: string
          required: true
          description: prohibited product id
        - in: body
          name: prohibited product
          description: a existing prohibited product.
          schema:
            type: object
            required:
              - name
              - detail
            properties:
              name:
                type: string
              detail:
                type: string
        - in: formData
          name: images
          description: a images prohibited product.
          type: file
      requestBody:
        content: 
          multipart/form-data: 
            schema:            
              type: object
              properties:      
                name:
                  type: string
                images:
                  type: string
                  format: base64 
                detail:
                  type: string
      responses:
        200:
          description: Update prohibited product successfully.
          schema:
            type: object
            example: {
                "name": "string",
                "detail": "string",
                "image": "path",
            }
        400:
          description: Update prohibited product failed.
      security:
        - bearerAuth: []
    delete:
      tags:
        - "prohibited-product"
      summary: "Delete a existing prohibited product"
      produces:
        - application/json
      parameters:
        - in: path
          name: id 
          type: string
          required: true
          description: prohibited product Id
      responses:
        200:
          description: Delete prohibited product successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
        400:
          description: Update prohibited product failed.
      security:
        - bearerAuth: []
...
  /product/{orderId}:
    post:
      tags:
        - "product"
      summary: "Customer add a new product to order"
      consumes:
        - application/json
      parameters:
        - in: path
          name: orderId
          type: string
          required: true
          description: orderId of order
        - in: body
          name: products
          description: array of products.
          schema:
            type: object      
            properties:
              products:
                type: array
                items:
                  type: object    
                  properties:              
                    name: 
                      type: string
                    quantity:
                      type: integer
                    unit: 
                      type: string   
                      enum: ['m3', 'kg', 'ton']
      
      responses:
        200:
          description: Add product(s) successfully.
        400:
          description: Add product(s) failed.
      security:
        - bearerAuth: []
    get:
      tags:
        - "product"
      summary: "Customer get product of an order"
      consumes:
        - application/json
      parameters:
        - in: path
          name: orderId
          type: string
          required: true
          description: orderId of order         
      responses:
        200:
          description: Get product successfully.
          schema:
            type: object
            example: {
                  "name": "string",
                  "quantity": "string",
                  "unit": "string",                 
            }
        400:
          description: Get product failed.
      security:
        - bearerAuth: []
  /product/{productId}:
    put:
      tags:
        - "product"
      summary: "Customer update product in waiting order"
      consumes:
        - application/json
      parameters:
        - in: path
          name: productId
          type: string
          required: true
          description: productId
        - in: body
          name: product
          description: updated product
          schema:
            type: object      
            properties:
              name: 
                type: string
              quantity:
                type: integer
              unit: 
                type: string   
                enum: ['m3', 'kg', 'ton']            
      responses:
        200:
          description: Update product successfully.
        400:
          description: Update product failed.
      security:
        - bearerAuth: []
    delete:
      tags:
        - "product"
      summary: "Customer delete product in waiting order"
      consumes:
        - application/json
      parameters:
        - in: path
          name: productId
          type: string
          required: true
          description: productId                      
      responses:
        200:
          description: Delete product successfully.
        400:
          description: Delete product failed.
      security:
        - bearerAuth: []
  /admin/product:
    get:
      tags:
        - "product"
      summary: "Admin get list of products"
      consumes:
        - application/json
      parameters:
        - in: query
          name: pageSize
          type: string
          description: limit number of product on a page
        - in: query
          name: page
          type: string
          description: num of page
        - in: query
          name: keyword
          type: string
          description: keyword
        - in: query
          name: sortBy
          type: string
          description: sort by (name, quantity, updatedAt)
        - in: query
          name: orderId
          type: string
          description:  orderId
      responses:
        200:
          description: Get product successfully.
          schema:
            type: object
            example: {
                  "name": "string",
                  "quantity": integer,
                  "unit": "string",                 
            }
        400:
          description: Get product failed.
      security:
        - bearerAuth: []
  /admin/product/{orderId}:
    post:
      tags:
        - "product"
      summary: "Admin add a new product to order"
      consumes:
        - application/json
      parameters:
        - in: path
          name: orderId
          type: string
          required: true
          description: orderId of order
        - in: body
          name: products
          description: array of products.
          schema:
            type: object      
            properties:
              products:
                type: array
                items:
                  type: object    
                  properties:              
                    name: 
                      type: string
                    quantity:
                      type: integer
                    unit: 
                      type: string   
                      enum: ['m3', 'kg', 'ton']
      
      responses:
        200:
          description: Add product(s) successfully.
        400:
          description: Add product(s) failed.
      security:
        - bearerAuth: []
  /admin/product/{productId}:
    put:
      tags:
        - "product"
      summary: "Admin update product in waiting order"
      consumes:
        - application/json
      parameters:
        - in: path
          name: productId
          type: string
          required: true
          description: productId
        - in: body
          name: product
          description: updated product
          schema:
            type: object      
            properties:
              name: 
                type: string
              quantity:
                type: integer
              unit: 
                type: string   
                enum: ['m3', 'kg', 'ton']            
      responses:
        200:
          description: Update product successfully.
        400:
          description: Update product failed.
      security:
        - bearerAuth: []
    delete:
      tags:
        - "product"
      summary: "Admin delete product in waiting order"
      consumes:
        - application/json
      parameters:
        - in: path
          name: productId
          type: string
          required: true
          description: productId                      
      responses:
        200:
          description: Delete product successfully.
        400:
          description: Delete product failed.
      security:
        - bearerAuth: []
...
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"